+++ workflows/review/lib/dispatch.ts
@@ -61,6 +61,7 @@ import {
     type Candidate,
     type Claim,
 } from "./dispatch-contracts";
+import {computeRoster, type RosterShed} from "./dispatch-roster";
 import {
     applyProvenanceGate,
     type DiffProvenance,
@@ -81,6 +82,13 @@ export {
     type ContractKind,
     type Verification,
 } from "./dispatch-contracts";
+export {
+    computeRoster,
+    DEFAULT_FINDERS,
+    SHED_RANKING,
+    type Roster,
+    type RosterShed,
+} from "./dispatch-roster";
 export {
     dedupeClaims,
     suppressOpenThreadDuplicates,
@@ -156,22 +164,6 @@ const DEFAULT_CONCURRENCY = 4;
 const TRIAGE = "pattern-triage";
 const RECONCILER = "thread-reconciler";
 const VALIDATOR = "claim-validator";
-const DEFAULT_FINDERS = ["correctness-reviewer", "skill-auditor"] as const;
-
-/**
- * The Step 3 dispatch/shed ranking, first-shed first. Fill order under the
- * invocation cap is this list reversed after the defaults and matched
- * lenses. An enabled reviewer this table does not know sheds after
- * `conventions` (generic before targeted).
- */
-const SHED_RANKING = [
-    "documentation",
-    "conventions",
-    "first-principles",
-    "holistic",
-    "completeness",
-    "test-adequacy",
-] as const;
 
 /* -------------------------------------------------------------------------- */
 /* Agent definitions (.claude/agents/<name>.md)                               */
@@ -227,79 +219,6 @@ const loadAgents = (
         }
     }
     return agents;
-};
-
-/* -------------------------------------------------------------------------- */
-/* Roster                                                                     */
-/* -------------------------------------------------------------------------- */
-
-export type RosterShed = {name: string; cause: "budget"};
-
-export type Roster = {
-    /** Finding producers to dispatch, in dispatch-ranking order. */
-    finders: string[];
-    /** Planned finding producers shed under the invocation cap. */
-    shed: RosterShed[];
-    /** Whether pattern-triage runs (full/scoped only). */
-    triage: boolean;
-    /** Whether the reconciler runs (staged threads exist). */
-    reconcile: boolean;
-};
-
-/**
- * Compute the dispatch roster per the staged depth plan and routing, capped
- * by `runBudget.maxReviewerInvocations` in the Step 3 dispatch ranking
- * (defaults, then matched lenses, then opt-ins by inverse shed order), every
- * capped-out entry recorded as a planned shed. Pipeline steps (triage,
- * reconciler, validator) never consume a slot.
- */
-export const computeRoster = (
-    depth: string,
-    routing: {
-        enabledReviewers?: unknown;
-        lensesToSpawn?: unknown;
-        runBudget?: {maxReviewerInvocations?: unknown};
-    },
-    hasThreads: boolean,
-): Roster => {
-    if (depth === "fast") {
-        return {finders: [], shed: [], triage: false, reconcile: hasThreads};
-    }
-    if (depth === "flip-gated") {
-        return {
-            finders: ["correctness-reviewer"],
-            shed: [],
-            triage: false,
-            reconcile: hasThreads,
-        };
-    }
-    const strings = (value: unknown): string[] =>
-        Array.isArray(value)
-            ? value.filter((v): v is string => typeof v === "string")
-            : [];
-    const lenses = strings(routing.lensesToSpawn);
-    const enabled = strings(routing.enabledReviewers).filter(
-        (name) => !lenses.includes(name),
-    );
-    const optIns = [...enabled].sort((a, b) => {
-        const rank = (name: string): number => {
-            const index = (SHED_RANKING as readonly string[]).indexOf(name);
-            return index === -1 ? -0.5 : index;
-        };
-        return rank(b) - rank(a);
-    });
-    const ranked = [...DEFAULT_FINDERS, ...lenses, ...optIns];
-
-    const capRaw = routing.runBudget?.maxReviewerInvocations;
-    const cap =
-        typeof capRaw === "number" && Number.isInteger(capRaw) && capRaw >= 0
-            ? capRaw
-            : ranked.length;
-    const finders = ranked.slice(0, Math.max(cap, DEFAULT_FINDERS.length));
-    const shed = ranked
-        .slice(finders.length)
-        .map((name): RosterShed => ({name, cause: "budget"}));
-    return {finders, shed, triage: true, reconcile: hasThreads};
 };
 
 /* -------------------------------------------------------------------------- */
